+++ src/Common.cpp
@@ -6,6 +6,14 @@
 #include <list>
 #include <algorithm>
 #include <fstream>
+#include <regex>
+#include <functional>
+#include <random>
+#include <cstdlib>
+
+#if !defined(OS_WIN)
+#include <unistd.h> // getcwd()
+#endif
 
 namespace cefpdf {
 
@@ -111,18 +119,18 @@ void parseCustomPageSize(PageSize& pageSize, const std::string& str)
                     pageSize.height = v;
                     break;
                 default:
-                    throw "Too many values in page size";
+                    throw std::string("Too many values in page size");
             }
 
             hasValue = true;
             value.erase();
         } else {
-            throw "Invalid character \"" + std::string(1, c) + "\" in page size";
+            throw std::string("Invalid character \"") + std::string(1, c) + std::string("\" in page size");
         }
     }
 
     if (!hasValue) {
-        throw "Invalid page size format: \"" + str + "\"";
+        throw std::string("Invalid page size format: \"") + str + std::string("\"");
     }
 }
 
@@ -176,18 +184,18 @@ void parseCustomPageMargin(PageMargin& pageMargin, const std::string& str)
                     pageMargin.left = v;
                     break;
                 default:
-                    throw "Too many values in margin";
+                    throw std::string("Too many values in margin");
             }
 
             hasValue = true;
             value.erase();
         } else {
-            throw "Invalid character \"" + std::string(1, c) + "\" in margin";
+            throw std::string("Invalid character \"") + std::string(1, c) + std::string("\" in margin");
         }
     }
 
     if (!hasValue) {
-        throw "Invalid margin format: \"" + str + "\"";
+        throw std::string("Invalid margin format: \"") + str + std::string("\"");
     }
 }
 
@@ -215,6 +223,104 @@ std::chrono::microseconds::rep microtime()
 {
     auto tt = std::chrono::system_clock::now().time_since_epoch();
     return std::chrono::duration_cast<std::chrono::microseconds>(tt).count();
+}
+
+std::string pathToUri(const std::string& path)
+{
+    std::string uri = path;
+
+#if defined(OS_WIN)
+    std::regex re("^[a-z]:\\\\", std::regex_constants::icase);
+    if (!std::regex_search(uri, re, std::regex_constants::match_continuous)) {
+        uri = getCurrentWorkingDirectory() + std::string("\\") + uri;
+    }
+
+    // De-windows the path
+    std::replace(uri.begin(), uri.end(), '\\', '/');
+    uri = std::string("/") + uri;
+#else
+    if (uri.front() != '/') {
+        uri = getCurrentWorkingDirectory() + std::string("/") + uri;
+    }
+#endif // OS_WIN
+
+    return std::string("file://") + uri;
+}
+
+std::string getCurrentWorkingDirectory()
+{
+#if defined(OS_WIN)
+    wchar_t lpBuffer[MAX_PATH];
+    ::GetCurrentDirectoryW(MAX_PATH, lpBuffer);
+    char result[2*MAX_PATH];
+    std::wcstombs(result, lpBuffer, sizeof(result));
+    return result;
+#else
+    char result[2*4096];
+    getcwd(result, 2*4096);
+    return result;
+#endif // OS_WIN
+}
+
+std::string getTempDirectory()
+{
+#if defined(OS_WIN)
+    wchar_t lpBuffer[MAX_PATH];
+    ::GetTempPathW(MAX_PATH, lpBuffer);
+    char result[2*MAX_PATH];
+    std::wcstombs(result, lpBuffer, sizeof(result));
+    return result;
+#else
+    const char* vars[4] = {"TMPDIR", "TMP", "TEMP", "TEMPDIR"};
+    for (int i = 0; i < 4; ++i) {
+        char* val = std::getenv(vars[i]);
+        if (val) {
+            return std::string(val) + "/";
+        }
+    }
+
+    return "/tmp/";
+#endif // OS_WIN
+}
+
+std::string reserveTempFile()
+{
+    std::string letters = "9876543210ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+    std::random_device rd;
+    std::mt19937 g(rd());
+    std::shuffle(letters.begin(), letters.end(), g);
+    std::string path = constants::temp + "cef_" + letters + ".pdf";
+
+    std::ifstream file;
+    file.open(path);
+    bool isGood = file.good();
+    file.close();
+
+    return isGood ? reserveTempFile() : path;
+}
+
+std::string loadTempFile(const std::string& path, bool remove)
+{
+    std::string content;
+    std::ifstream output;
+
+    output.open(path, std::ifstream::binary);
+    if (output.good()) {
+        content.assign((std::istreambuf_iterator<char>(output)), std::istreambuf_iterator<char>());
+        output.close();
+        if (remove) {
+            deleteTempFile(path);
+        }
+
+        return content;
+    }
+
+    throw std::string("Cannot open file: \"") + path + std::string("\"");
+}
+
+bool deleteTempFile(const std::string& path)
+{
+    return 0 == std::remove(path.c_str());
 }
 
 } // namespace cefpdf